+++ DesignerMacros/src/services-cqrs-crud/create-crud-macro-advanced-mapping/create-crud-macro-advanced-mapping.ts
@@ -9,11 +9,11 @@ async function execute(element: IElementApi, domainClass?: IElementApi) {
     }
 
     if (privateSettersOnly && !hasConstructor(entity)) {
-        await dialogService.error(
-`Unable to create CQRS Operations.
+        await dialogService.warn(
+`Partial CQRS Operation Creation.
+Some CQRS operations were created successfully, but was limited due to private setters being enabled, and no constructor is present for entity '${entity.getName()}'.
 
-Private Setters are enabled with no constructor present on entity '${entity.getName()}'. In order to create CQRS Operations for that entity, either disable private setters or model a constructor element and try again.`);
-        return;
+To avoid this limitation in the future, either disable private setters or add a constructor element to the entity.`);
     }
 
     const owningEntity = DomainHelper.getOwningAggregate(entity);
@@ -24,7 +24,9 @@ Private Setters are enabled with no constructor present on entity '${entity.getN
 
     const resultDto = cqrsCrud.createCqrsResultTypeDto(entity, folder);
 
-    convertToAdvancedMapping.convertCommand(cqrsCrud.createCqrsCreateCommand(entity, folder, primaryKeys));
+    if (!privateSettersOnly || hasConstructor(entity)) {
+        convertToAdvancedMapping.convertCommand(cqrsCrud.createCqrsCreateCommand(entity, folder, primaryKeys));
+    }
 
     if (hasPrimaryKey) {
         convertToAdvancedMapping.convertQuery(cqrsCrud.createCqrsFindByIdQuery(entity, folder, resultDto));